+++ components/utilities/direction/private/language-direction.jsx
@@ -1,13 +1,15 @@
 import React, { Component } from 'react';
 
-import DirectionSettings from './direction-settings';
+// eslint-disable-next-line camelcase
+import UNSAFE_DirectionSettings from '..';
 
 const LanguageDirectionHOC = (WrappedComponent) => {
 	const componentName =
 		WrappedComponent.displayName || WrappedComponent.name || 'Component';
 	return class LanguageDirection extends Component {
 		static displayName = `LanguageDirection(${componentName})`;
-		static contextType = DirectionSettings;
+		// eslint-disable-next-line camelcase
+		static contextType = UNSAFE_DirectionSettings;
 
 		render() {
 			const props = { ...this.props };
